refactor resource. wip content

+++ resource/content/manager.go
@@ -4,24 +4,33 @@ import (
 	"context"
 	"distudio.com/mage/model"
 	"distudio.com/page/resource"
+	"distudio.com/page/resource/attachment"
+	"distudio.com/page/resource/identity"
+	"errors"
 	"google.golang.org/appengine/log"
+	"reflect"
+	"sort"
 )
 
-type Manager struct {}
+type Manager struct{}
 
 func (manager Manager) NewResource(ctx context.Context) (resource.Resource, error) {
 	return &Content{}, nil
 }
 
-
 func (manager Manager) FromId(ctx context.Context, id string) (resource.Resource, error) {
+	current, _ := ctx.Value(identity.KeyUser).(identity.User)
+	if !current.HasPermission(identity.PermissionReadContent) {
+		return nil, resource.NewPermissionError(identity.PermissionReadContent)
+	}
+
 	cont := Content{}
 	if err := model.FromStringID(ctx, &cont, id, nil); err != nil {
 		log.Errorf(ctx, "could not retrieve content %s: %s", id, err.Error())
 		return nil, err
 	}
 
-	q := model.NewQuery((*Attachment)(nil))
+	q := model.NewQuery((*attachment.Attachment)(nil))
 	q = q.WithField("Parent =", cont.Slug)
 	if err := q.GetMulti(ctx, &cont.Attachments); err != nil {
 		log.Errorf(ctx, "could not retrieve content %s attachments: %s", id, err.Error())
@@ -31,7 +40,12 @@ func (manager Manager) FromId(ctx context.Context, id string) (resource.Resource
 	return &cont, nil
 }
 
-func (manager Manager) ListOf(ctx context.Context, opts resource.ListOptions) ([]resource.Resource, error) {
+func (manager Manager) ListOf(ctx context.Context, opts resource.ListOptions) ([]interface{}, error) {
+	current, _ := ctx.Value(identity.KeyUser).(identity.User)
+	if !current.HasPermission(identity.PermissionReadContent) {
+		return nil, resource.NewPermissionError(identity.PermissionReadContent)
+	}
+
 	var conts []*Content
 	q := model.NewQuery(&Content{})
 	q = q.OffsetBy(opts.Page * opts.Size)
@@ -51,7 +65,7 @@ func (manager Manager) ListOf(ctx context.Context, opts resource.ListOptions) ([
 		return nil, err
 	}
 
-	resources := make([]resource.Resource, len(conts))
+	resources := make([]interface{}, len(conts))
 	for i := range conts {
 		resources[i] = resource.Resource(conts[i])
 	}
@@ -59,7 +73,57 @@ func (manager Manager) ListOf(ctx context.Context, opts resource.ListOptions) ([
 	return resources, nil
 }
 
+func (manager Manager) ListOfProperties(ctx context.Context, opts resource.ListOptions) ([]interface{}, error) {
+	current, _ := ctx.Value(identity.KeyUser).(identity.User)
+	if !current.HasPermission(identity.PermissionReadContent) {
+		return nil, resource.NewPermissionError(identity.PermissionReadContent)
+	}
+
+	a := []string{"category", "topic", "name"} // list property accepted
+	name := opts.Property
+
+	i := sort.Search(len(a), func(i int) bool { return name <= a[i] })
+	if i < len(a) && a[i] == name {
+		// found
+	} else {
+		return nil, errors.New("no property found")
+	}
+
+	var conts []*Content
+	q := model.NewQuery(&Content{})
+	q = q.OffsetBy(opts.Page * opts.Size)
+
+	if opts.Order != "" {
+		dir := model.ASC
+		if opts.Descending {
+			dir = model.DESC
+		}
+		q = q.OrderBy(opts.Order, dir)
+	}
+
+	q = q.Distinct(name)
+	q = q.Limit(opts.Size + 1)
+	err := q.GetAll(ctx, &conts)
+	if err != nil {
+		log.Errorf(ctx, "Error retrieving result: %+v", err)
+		return nil, err
+	}
+	var result []interface{}
+	for _, c := range conts {
+		value := reflect.ValueOf(c).Elem().FieldByName(name).String()
+		if len(value) > 0 {
+			result = append(result, value)
+		}
+	}
+	return result, nil
+}
+
 func (manager Manager) Save(ctx context.Context, res resource.Resource) error {
+	current, _ := ctx.Value(identity.KeyUser).(identity.User)
+	if !current.HasPermission(identity.PermissionEditContent) {
+		return resource.NewPermissionError(identity.PermissionEditContent)
+	}
+
 	content := res.(*Content)
 	// input is valid, create the resource
 	opts := model.CreateOptions{}
@@ -81,8 +145,13 @@ func (manager Manager) Save(ctx context.Context, res resource.Resource) error {
 	return nil
 }
 
-func (manager Manager) Delete(ctx context.Context, resource resource.Resource) error {
-	content := resource.(*Content)
+func (manager Manager) Delete(ctx context.Context, res resource.Resource) error {
+	current, _ := ctx.Value(identity.KeyUser).(identity.User)
+	if !current.HasPermission(identity.PermissionEditContent) {
+		return resource.NewPermissionError(identity.PermissionEditContent)
+	}
+
+	content := res.(*Content)
 	err := model.Delete(ctx, content, nil)
 	if err != nil {
 		log.Errorf(ctx, "error deleting content %s: %s", content.Slug, err.Error())
@@ -90,8 +159,8 @@ func (manager Manager) Delete(ctx context.Context, resource resource.Resource) e
 	}
 
 	// delete attachments with parent = slug
-	attachments := make([]*Attachment, 0, 0)
-	q := model.NewQuery(&Attachment{})
+	attachments := make([]*attachment.Attachment, 0, 0)
+	q := model.NewQuery(&attachment.Attachment{})
 	q.WithField("Parent =", content.Slug)
 	err = q.GetMulti(ctx, &attachments)
 	if err != nil {
@@ -102,15 +171,10 @@ func (manager Manager) Delete(ctx context.Context, resource resource.Resource) e
 	for _, attachment := range attachments {
 		err = model.Delete(ctx, attachment, nil)
 		if err != nil {
-			log.Errorf(ctx,"error deleting attachment %+v: %s", attachment, err.Error())
+			log.Errorf(ctx, "error deleting attachment %+v: %s", attachment, err.Error())
 			return err
 		}
 	}
 
 	return nil
 }
-
-func (manager Manager) PropertyValues(ctx context.Context, properties []string) ([]string, error) {
-
-}
-
